linefix checks for non-fried images

+++ line/linefix.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
 dir=$(cd -P -- "$(dirname -- "${BASH_SOURCE[0]}")" && pwd -P)
 for file in "$1"/*; do
-  if [[ $file == *.meta ]]; then
-    continue
-  fi
-  $dir/pngdefry -so $file
-  rm $file
+	if [[ $file == *.meta ]]; then
+		continue
+	fi
+	if [[ "$($dir/pngfix $file)" =~ "CgBI" ]]; then
+		$dir/pngdefry -so $file
+		rm $file
+	fi
 done
 echo "All broken files have been reprocessed and deleted."
